Refactor the cmd/ folder layout
Those changes aims to follow the advices from https://go.dev/doc/modules/layout
