+++ cmd/server/commands/bootstrap.go
@@ -1,4 +1,4 @@
-package server
+package commands
 
 import (
 	"fmt"
