+++ cmd/server/commands/root.go
@@ -1,4 +1,4 @@
-package server
+package commands
 
 import "github.com/spf13/cobra"
