+++ cmd/server/main.go
@@ -1,17 +1,15 @@
-package cmd
+package main
 
 import (
 	"os"
 
 	"github.com/spf13/cobra"
-	"github.com/theduckcompany/duckcloud/cmd/server"
+	"github.com/theduckcompany/duckcloud/cmd/server/commands"
 )
 
 const binaryName = "duckcloud"
 
-// Execute adds all child commands to the root command and sets flags appropriately.
-// This is called by main.main(). It only needs to happen once to the rootCmd.
-func Execute() {
+func main() {
 	cmd := &cobra.Command{
 		Use:   binaryName,
 		Short: "Manage your duckcloud instance in your terminal.",
@@ -22,7 +20,7 @@ func Execute() {
 	// tb := toolbox.NewProd()
 
 	// Subcommands
-	cmd.AddCommand(server.NewServerCmd(binaryName))
+	cmd.AddCommand(commands.NewServerCmd(binaryName))
 
 	err := cmd.Execute()
 	if err != nil {
